Factor out data only ophys / behavior + ophys methods
This commit factors out 'data only' (LIMS/BehaviorOphysJson/NWB) API
methods from methods that are required by the OphysSession class.

This commit creates a new set of classes (BehaviorDataXforms, BehaviorOphysDataXforms)
that transforms data obtained from 'data only' methods into sources for
the OphysSession class.
