+++ allensdk/brain_observatory/behavior/session_apis/data_transforms/__init__.py
@@ -1,0 +1,2 @@
+from allensdk.brain_observatory.behavior.session_apis.data_transforms.behavior_data_xforms import BehaviorDataXforms  # noqa: F401, E501
+from allensdk.brain_observatory.behavior.session_apis.data_transforms.behavior_ophys_data_xforms import BehaviorOphysDataXforms  # noqa: F401, E501
