+++ allensdk/test/brain_observatory/behavior/test_behavior_data_lims_api.py
@@ -15,8 +15,8 @@ from allensdk.brain_observatory.running_speed import RunningSpeed
 from allensdk.core.exceptions import DataFrameIndexError
 
 mock_db_credentials = DbCredentials(dbname='mock_db', user='mock_user',
-                                      host='mock_host', port='mock_port',
-                                      password='mock')
+                                    host='mock_host', port='mock_port',
+                                    password='mock')
 
 
 @pytest.fixture
@@ -31,9 +31,13 @@ def MockBehaviorLimsApi():
             super().__init__(behavior_session_id=8675309,
                              lims_credentials=mock_db_credentials,
                              mtrain_credentials=mock_db_credentials)
+            self.foraging_id = 123456
 
         def _get_ids(self):
             return {}
+
+        def get_behavior_stimulus_file(self):
+            return "dummy_stimulus_file.pkl"
 
         def _behavior_stimulus_file(self):
             data = {
@@ -80,7 +84,7 @@ def MockBehaviorLimsApi():
 def MockApiRunSpeedExpectedError():
     class MockApiRunSpeedExpectedError(BehaviorLimsApi):
         """
-        Mock class that overrides some functions to provide test data and 
+        Mock class that overrides some functions to provide test data and
         initialize without calls to db.
         """
         def __init__(self):
@@ -169,7 +173,7 @@ class TestBehaviorRegression:
     sources (and I'm not sure how it's uploaded in the database).
 
     Do not test `get_licks` regression because the licks come from two
-    different sources and are recorded differently (behavior pickle file in 
+    different sources and are recorded differently (behavior pickle file in
     BehaviorLimsApi; sync file in BehaviorOphysLimeApi)
     """
     @classmethod
